(GH-373) Convert Puppet Resource Command to Feature
